refactor: modularize exporter
